Read Attributes.

+++ src/HDF5.NET/FileFormat/Level2/Level2A2/Level2A2d/FloatingPointPropertyDescription.cs
@@ -14,7 +14,7 @@ namespace HDF5.NET
             this.ExponentSize = reader.ReadByte();
             this.MantissaLocation = reader.ReadByte();
             this.MantissaSize = reader.ReadByte();
-            this.ExponentBias = reader.ReadUInt16();
+            this.ExponentBias = reader.ReadUInt32();
         }
 
         #endregion
